Moves the template downloader, and manipulator into it's own class
